refactor:challenge type 분리
